refactor(core): normalize architecture and resolve system dependencies for pipeline 3
- 환경 관리자(EnvironmentManager)를 신설하여 보드 설정 시 타겟 월드만 최적화하도록 서버 전역 오염 위험 차단
- 메인 루프를 각 매니저(Timer, Scoreboard)의 자율적 관찰(Polling) 기반으로 분리하여 순환 참조 원천 해결 및 안정성 확보
- 전투 및 가이드 로직을 전담 시스템(CombatManager, BoardVisualManager)으로 이관하여 입력 컨트롤러(리스너)의 책임을 대폭 압축- 외부 플러그인(MythicMobs) 설정 로직을 도메인 패키지로 이관하고, onEnable 부트스트래핑 순서를 최적화하여 시스템 응집도 향상
