+++ ARCHITECTURE.md
@@ -24,8 +24,9 @@
     * 게임의 근간이 되는 거대한 물리적 규칙이나 수학적 뼈대를 가집니다. 그 자체가 하나의 하위 시스템이므로 명시적인 생성자로 생성합니다.
 3. **관리자 및 시스템 (Manager / System)**
     * 데이터와 코어 객체들을 조립하여 게임의 시간과 상태 흐름을 전역적으로 통제합니다. 복잡한 비즈니스 로직이 구현되는 곳입니다.
-4. **입력 컨트롤러 (Input Controller / Command)**
+4. **입력 컨트롤러 (Command)**
     * 유저의 입력을 감지하는 가장 바깥쪽 껍데기입니다. 비즈니스 로직을 직접 처리하지 않고 관리자에게 제어권을 토스합니다.
+    * 일반 플레이어용 커맨드는 반드시 게임 비즈니스 규칙을 준수해야 하며, 시스템 계층의 검증 로직을 통과해야 합니다.
 
 ### 2.2. Engine-Native
 * 외부 기술로부터 도메인을 격리하려는 시도를 지양합니다.
@@ -58,6 +59,11 @@
 ### 3.3. 의존성 관리 전략
 *   **Bukkit/Paper API:** 1급 객체로 직접 수용합니다. 매핑 오버헤드 없이 도메인 로직 내에서 직접 사용합니다.
 *   **외부 라이브러리 및 플러그인:** ACF, FastBoard, MythicMobs 등 변동성이 높은 외부 의존성은 반드시 해당 기능을 담당하는 Feature 패키지 내에서 격리하여 사용합니다.
+
+### 3.4. 관리자 명령어 예외 규정
+* **관리자 명령어의 특수성:** 관리자 전용 명령어는 시스템 장애 복구 및 강제 개입을 목적으로 하므로, 일반적인 의존성 최소화 원칙의 예외를 적용합니다.
+* **다중 의존성 허용:** 문제를 해결하기 위해 보드, 기물, 타이머 등 게임 내 모든 서브시스템 매니저를 동시에 주입받는 것을 허용합니다.
+* **물리적 격리:** 다른 도메인 로직의 아키텍처 오염을 방지하기 위해 반드시 admin 패키지에 격리하여 관리합니다.
 
 ---
