+++ src/main/java/dev/tecte/chesswar/ChessWar.java
@@ -1,16 +1,22 @@
 package dev.tecte.chesswar;
 
 import co.aikar.commands.PaperCommandManager;
+import dev.tecte.chesswar.admin.AdminCommand;
 import dev.tecte.chesswar.board.BoardBlockListener;
-import dev.tecte.chesswar.board.BoardCommand;
 import dev.tecte.chesswar.board.BoardManager;
 import dev.tecte.chesswar.board.BoardVisualGuideListener;
+import dev.tecte.chesswar.board.BoardVisualManager;
 import dev.tecte.chesswar.board.MoveValidator;
+import dev.tecte.chesswar.game.CombatManager;
+import dev.tecte.chesswar.game.EnvironmentManager;
+import dev.tecte.chesswar.game.GameCommand;
 import dev.tecte.chesswar.game.GameEntityListener;
 import dev.tecte.chesswar.game.GameManager;
 import dev.tecte.chesswar.game.GameReadyListener;
 import dev.tecte.chesswar.game.ScoreboardManager;
+import dev.tecte.chesswar.game.StatsCommand;
 import dev.tecte.chesswar.game.TimerManager;
+import dev.tecte.chesswar.piece.PieceCommand;
 import dev.tecte.chesswar.piece.PieceDamageListener;
 import dev.tecte.chesswar.piece.PieceInteractListener;
 import dev.tecte.chesswar.piece.PieceManager;
@@ -18,19 +24,8 @@ import dev.tecte.chesswar.piece.PieceSelectionListener;
 import lombok.Getter;
 import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
-import org.bukkit.Bukkit;
-import org.bukkit.Difficulty;
-import org.bukkit.GameRule;
-import org.bukkit.World;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
-import org.bukkit.scheduler.BukkitRunnable;
-
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
 
 @Getter
 @Accessors(fluent = true)
@@ -38,20 +33,21 @@ import java.nio.file.Files;
 public class ChessWar extends JavaPlugin {
     private BoardManager boardManager;
     private PieceManager pieceManager;
+    private EnvironmentManager environmentManager;
     private GameManager gameManager;
-    private TimerManager timerManager;
-    private ScoreboardManager scoreboardManager;
     private MoveValidator moveValidator;
+    private BoardVisualManager boardVisualManager;
+    private TimerManager timerManager;
+    private CombatManager combatManager;
+    private ScoreboardManager scoreboardManager;
 
     @Override
     public void onEnable() {
         initializeManagers();
-        registerCommands();
+        pieceManager.setupMythicMobs(this);
         registerListeners();
-        startHeartbeat();
-        setupMythicMobs();
-        setupOptimalEngineSettings();
-        log.info("ChessWar has been enabled with optimized engine settings!");
+        registerCommands();
+        log.info("ChessWar has been enabled!");
     }
 
     @Override
@@ -62,123 +58,58 @@ public class ChessWar extends JavaPlugin {
     private void initializeManagers() {
         boardManager = new BoardManager();
         pieceManager = new PieceManager();
+        environmentManager = new EnvironmentManager();
         gameManager = new GameManager();
-        timerManager = new TimerManager(this, gameManager, pieceManager);
-        scoreboardManager = new ScoreboardManager(gameManager, timerManager);
         moveValidator = new MoveValidator(gameManager, pieceManager);
-    }
-
-    private void startHeartbeat() {
-        new BukkitRunnable() {
-            @Override
-            public void run() {
-                timerManager.tick();
-                scoreboardManager.updateAll();
-            }
-        }.runTaskTimer(this, 0L, 20L);
-    }
-
-    private void registerCommands() {
-        PaperCommandManager commandManager = new PaperCommandManager(this);
-
-        commandManager.registerCommand(new BoardCommand(
+        boardVisualManager = new BoardVisualManager(gameManager, boardManager, pieceManager, moveValidator);
+        timerManager = new TimerManager(this, gameManager, pieceManager);
+        combatManager = new CombatManager(
                 this,
                 gameManager,
                 boardManager,
                 pieceManager,
-                timerManager,
-                moveValidator
-        ));
+                moveValidator,
+                timerManager
+        );
+        scoreboardManager = new ScoreboardManager(this, gameManager, timerManager);
     }
 
     private void registerListeners() {
         PluginManager pluginManager = getServer().getPluginManager();
 
+        pluginManager.registerEvents(new PieceSelectionListener(this, gameManager), this);
         pluginManager.registerEvents(
-                new PieceInteractListener(this, gameManager, boardManager, pieceManager, moveValidator),
+                new GameReadyListener(this, gameManager, boardManager, timerManager),
                 this
         );
-        pluginManager.registerEvents(
-                new PieceDamageListener(this, gameManager, boardManager, pieceManager, moveValidator, timerManager),
-                this
-        );
-        pluginManager.registerEvents(
-                new BoardVisualGuideListener(gameManager, boardManager, pieceManager, moveValidator),
-                this
-        );
-        pluginManager.registerEvents(
-                new GameEntityListener(gameManager, pieceManager),
-                this
-        );
-        pluginManager.registerEvents(
-                new PieceSelectionListener(this, gameManager, timerManager),
-                this
-        );
-        pluginManager.registerEvents(new GameReadyListener(this, gameManager, boardManager, timerManager), this);
+        pluginManager.registerEvents(new PieceInteractListener(gameManager, combatManager), this);
+        pluginManager.registerEvents(new PieceDamageListener(combatManager), this);
+        pluginManager.registerEvents(new BoardVisualGuideListener(boardVisualManager), this);
         pluginManager.registerEvents(new BoardBlockListener(gameManager, boardManager), this);
+        pluginManager.registerEvents(new GameEntityListener(gameManager, pieceManager), this);
         pluginManager.registerEvents(timerManager, this);
     }
 
-    private void setupMythicMobs() {
-        Plugin mythicMobs = getServer().getPluginManager().getPlugin("MythicMobs");
+    private void registerCommands() {
+        PaperCommandManager commandManager = new PaperCommandManager(this);
 
-        if (mythicMobs == null) {
-            return;
-        }
-
-        File targetFile = new File(new File(mythicMobs.getDataFolder(), "Mobs"), "Piece.yml");
-
-        if (targetFile.exists()) {
-            return;
-        }
-
-        File parentDir = targetFile.getParentFile();
-
-        if (parentDir != null && !parentDir.exists() && !parentDir.mkdirs()) {
-            log.error("Failed to create MythicMobs piece directory.");
-            return;
-        }
-
-        try (InputStream in = getResource("mobs/Piece.yml")) {
-            if (in != null) {
-                Files.copy(in, targetFile.toPath());
-                log.info("Successfully synced Piece.yml to MythicMobs folder.");
-            }
-        } catch (IOException e) {
-            log.error("Failed to sync Piece.yml: {}", e.getMessage());
-        }
-    }
-
-    private void setupOptimalEngineSettings() {
-        Bukkit.getWorlds().forEach(world -> {
-            setGameRuleIfDifferent(world, GameRule.NATURAL_REGENERATION, false);
-            setGameRuleIfDifferent(world, GameRule.DO_DAYLIGHT_CYCLE, false);
-            setGameRuleIfDifferent(world, GameRule.DO_WEATHER_CYCLE, false);
-            setGameRuleIfDifferent(world, GameRule.DO_MOB_SPAWNING, false);
-            setGameRuleIfDifferent(world, GameRule.DO_TRADER_SPAWNING, false);
-            setGameRuleIfDifferent(world, GameRule.DO_PATROL_SPAWNING, false);
-            setGameRuleIfDifferent(world, GameRule.MOB_GRIEFING, false);
-            setGameRuleIfDifferent(world, GameRule.KEEP_INVENTORY, true);
-            setGameRuleIfDifferent(world, GameRule.DO_FIRE_TICK, false);
-            setGameRuleIfDifferent(world, GameRule.DO_TILE_DROPS, false);
-            setGameRuleIfDifferent(world, GameRule.ANNOUNCE_ADVANCEMENTS, false);
-            setGameRuleIfDifferent(world, GameRule.SPAWN_RADIUS, 0);
-
-            if (world.getDifficulty() != Difficulty.NORMAL) {
-                world.setDifficulty(Difficulty.NORMAL);
-            }
-
-            if (!world.getPVP()) {
-                world.setPVP(true);
-            }
-        });
-    }
-
-    private <T> void setGameRuleIfDifferent(World world, GameRule<T> rule, T value) {
-        T currentValue = world.getGameRuleValue(rule);
-
-        if (currentValue != null && !currentValue.equals(value)) {
-            world.setGameRule(rule, value);
-        }
+        commandManager.registerCommand(new GameCommand(
+                this,
+                gameManager,
+                boardManager,
+                pieceManager,
+                timerManager
+        ));
+        commandManager.registerCommand(new PieceCommand(gameManager, pieceManager, timerManager));
+        commandManager.registerCommand(new StatsCommand(gameManager));
+        commandManager.registerCommand(new AdminCommand(
+                this,
+                gameManager,
+                boardManager,
+                pieceManager,
+                moveValidator,
+                combatManager,
+                environmentManager
+        ));
     }
 }
